\name{getZipDir}
\alias{getZipDir}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{getZipDir}
\description{
Get the ZIP archive for testing.
}
\usage{
getZipDir()
}
%- maybe also 'usage' for other objects documented here.
\details{
Starts by checking the environmental variable GS_ZIP_ARCHIVE.
Then checks a couple of standard paths from the development environment.
}
\value{
A path that exists or NULL if no paths are found.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
