\name{ReadMatrixRcpp-package}
\alias{ReadMatrixRcpp-package}
\alias{ReadMatrixRcpp}
\docType{package}
\title{
  Read large matrix files using C++
}
\description{
  Read large matrix files using C++
}
\details{
  Read large matrix files using C++
}
\author{
}
\references{
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{ package }
\seealso{
}
\examples{
}
